Refactor processing, get rid of strategies.py
We decided that we don't want to implement the strategy abstraction right now.
Moved search chunk related functionality into finder.py and deleted strategies.py
The functions are ordered in the call order.
